♻️refactor: 체험 상세 페이지를 CSR 방식으로 변경 및 페이지 반영 - API 연동을 위한 타입 추가 - 반응형 대응을 위한 useResponsive 훅 사용
